chore: fix some TODO
