+++ src/main/java/dev/tonimatas/listeners/RouletteListener.java
@@ -3,10 +3,8 @@ package dev.tonimatas.listeners;
 import dev.tonimatas.roulette.bets.*;
 import dev.tonimatas.tasks.RouletteTask;
 import dev.tonimatas.util.Messages;
-import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Member;
-import net.dv8tion.jda.api.entities.MessageEmbed;
 import net.dv8tion.jda.api.events.interaction.command.CommandAutoCompleteInteractionEvent;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
@@ -14,7 +12,6 @@ import net.dv8tion.jda.api.interactions.commands.Command;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Stream;
 
 public class RouletteListener extends ListenerAdapter {
@@ -65,22 +62,6 @@ public class RouletteListener extends ListenerAdapter {
                     event.reply("Invalid bet option \"" + option + "\" for \"" + type + "\".").setEphemeral(true).queue(Messages.deleteBeforeX());
                 }
             }
-
-            // TODO: Move to new BankListener
-            case "money" -> {
-                long money = rouletteTask.getMoney(member.getId());
-                event.reply("You have " + money + "€.").queue(Messages.deleteBeforeX());
-            }
-
-            // TODO: The same as above
-            case "money-top" -> {
-                MessageEmbed embed = new EmbedBuilder()
-                        .setTitle("Money Top")
-                        .setDescription(getMoneyTopString(guild))
-                        .build();
-                
-                event.replyEmbeds(embed).queue(Messages.deleteBeforeX());
-            }
         }
     }
 
@@ -109,38 +90,7 @@ public class RouletteListener extends ListenerAdapter {
             }
         }
     }
-    
-    // TODO: Don't print empty string. 1. Unknown \n 2. Unknown...
-    private String getMoneyTopString(Guild guild) {
-        List<Map.Entry<String, Long>> sortedList = rouletteTask.getBank().entrySet()
-                .stream()
-                .sorted((a, b) ->
-                        b.getValue().compareTo(a.getValue()))
-                .toList();
 
-        int counter = 0;
-        StringBuilder text = new StringBuilder();
-
-        for (Map.Entry<String, Long> entry : sortedList) {
-            if (counter >= 5) break;
-
-            Member member = guild.getMemberById(entry.getKey());
-
-            String name = (member != null) ? member.getEffectiveName() : "Unknown";
-
-            text.append((counter + 1))
-                    .append(". ")
-                    .append(name)
-                    .append(" ")
-                    .append(entry.getValue())
-                    .append("€\n");
-
-            counter++;
-        }
-        
-        return text.toString();
-    }
-    
     private Bet getBet(String type, String id, String option, long money) {
         return switch (type) {
             case "color" -> new ColorBet(id, option, money);
